requirements.txt
Django==2.1
django-jet==1.0.7
django-ckeditor==5.6.1
django-constance==2.2.0
django-debug-toolbar==1.9.1
django-imagekit==4.0.2
django-rosetta==0.8.3
django-modeltranslation==0.12.2
django-picklefield==1.0.0
Unipath==1.1
python-memcached==1.59
feedparser==5.2.1
celery==4.3.0
django-environ==0.4.5
django-celery-results==1.1.2
django-hijack-admin==2.1.10
mysqlclient==1.4.6
django-structlog==1.3.3
pandas==0.24.2
pandas-datareader==0.8.1
TA-Lib==0.4.17
plotly==4.2.1
scipy==1.3.3
finta==0.3.9
cython==0.29.14
sklearn==0.0
scikit-learn==0.21.3
matplotlib==3.0.3
